🧪 test: add file test script

+++ file/file/file.py
@@ -1,24 +1,53 @@
 """File"""
 
+import os
+
 from session import session
 
 class File:
     """File"""
 
-    def __init__(self, source, work_session):
+    def __init__(self, scope, source, path, work_session):
+        self.scope = scope
         self.source = source
+        self.path = path
         self.session = work_session
 
     def filter_file(self, param=None):
+        if self.source:
+            if not param:
+                param = {'fileSource': self.source}
+            else:
+                param['fileSource'] = self.source
+        if self.scope:
+            if not param:
+                param = {'fileScope': self.scope}
+            else:
+                param['fileScope'] = self.scope
+
+        val = self.session.get('/api/v1/static/file/filter/', param)
+        if val and val['errorCode'] == 0:
+            return val['values']
+        print('--------filter_file-----------')
+        print(val['reason'])
+        return None
+
+    def query_file(self, file_id, param = None):
         if not param:
             param = {'fileSource': self.source}
         else:
             param['fileSource'] = self.source
+        if self.scope:
+            if not param:
+                param = {'fileScope': self.scope}
+            else:
+                param['fileScope'] = self.scope
 
-        val = self.session.get('/static/file/query/', param)
+        val = self.session.get('/api/v1/static/file/query/{0}'.format(file_id), param)
         if val and val['errorCode'] == 0:
-            return val['file']
-        print('--------filter_file-----------')
+            return val['value']
+
+        print('--------query_file-----------')
         print(val['reason'])
         return None
 
@@ -27,21 +56,44 @@ class File:
             param = {'fileSource': self.source}
         else:
             param['fileSource'] = self.source
+        if self.scope:
+            if not param:
+                param = {'fileScope': self.scope}
+            else:
+                param['fileScope'] = self.scope
 
-        val = self.session.delete('/static/file/delete/{0}'.format(file_id), param)
+        val = self.session.delete('/api/v1/static/file/delete/{0}'.format(file_id), param)
         if val and val['errorCode'] == 0:
-            return val['file']
+            return val['value']
 
         print('--------delete_file-----------')
         print(val['reason'])
         return None
 
-    def upload_file(self, file_path):
-        params = {'fileSource': self.source, 'key-name': 'file'}
+    def upload_file(self, file_path, params = None):
+        if not params:
+            params = {'fileSource': self.source}
+        else:
+            params['fileSource'] = self.source
+        if self.scope:
+            if not params:
+                params = {'fileScope': self.scope}
+            else:
+                params['fileScope'] = self.scope
+        if self.path:
+            if not params:
+                params = {'filePath': self.path}
+            else:
+                params['filePath'] = self.path
+        if not params:
+            params = {'key-name': 'file'}
+        else:
+            params['key-name'] = 'file'
+
         files = {'file': open(file_path, 'rb')}
-        val = self.session.upload('/static/file/upload/', files=files, params=params)
+        val = self.session.upload('/static/file/', files=files, params=params)
         if val and val['errorCode'] == 0:
-            return val['file']
+            return val['value']
         print('--------upload_file-----------')
         print(val['reason'])
         return None
@@ -59,14 +111,13 @@ class File:
 def main(server_url, namespace):
     """main"""
     work_session = session.MagicSession('{0}'.format(server_url), namespace)
-    app = File("autotest", work_session)
+    app = File("abc","xyz", None, work_session)
     file_path = "./file/file.py"
     new_file = app.upload_file(file_path)
     if not new_file:
         print('upload file failed')
         return False
 
-    print(new_file)
     file_list = app.filter_file()
     if not file_list or len(file_list) < 0:
         print('filter file failed')
@@ -76,6 +127,15 @@ def main(server_url, namespace):
     file_val = app.download_file(new_file['token'], new_file_path)
     if not file_val:
         print('download file failed')
+        return False
+    os.remove(new_file_path)
+
+    cur_file = app.query_file(new_file['id'])
+    if not cur_file:
+        print('query file failed')
+        return False
+    if cur_file['token'] != new_file['token']:
+        print('query file failed, mismatch file')
         return False
 
     old_file = app.delete_file(new_file['id'])
